fix(frontend): type shared quick action workflow
